+++ mopidy/backends/spotify/playback.py
@@ -5,6 +5,7 @@ import functools
 
 from spotify import Link, SpotifyError
 
+from mopidy import audio
 from mopidy.backends import base
 
 
@@ -30,6 +31,10 @@ class SpotifyPlaybackProvider(base.BasePlaybackProvider):
         'width=(int)16, depth=(int)16, signed=(boolean)true, '
         'rate=(int)44100')
 
+    def __init__(self, *args, **kwargs):
+        super(SpotifyPlaybackProvider, self).__init__(*args, **kwargs)
+        self._first_seek = False
+
     def play(self, track):
         if track.uri is None:
             return False
@@ -42,10 +47,13 @@ class SpotifyPlaybackProvider(base.BasePlaybackProvider):
         seek_data_callback_bound = functools.partial(
             seek_data_callback, spotify_backend)
 
+        self._first_seek = True
+
         try:
             self.backend.spotify.session.load(
                 Link.from_string(track.uri).as_track())
             self.backend.spotify.session.play(1)
+            self.backend.spotify.buffer_timestamp = 0
 
             self.audio.prepare_change()
             self.audio.set_appsrc(
@@ -75,5 +83,12 @@ class SpotifyPlaybackProvider(base.BasePlaybackProvider):
 
     def on_seek_data(self, time_position):
         logger.debug('playback.on_seek_data(%d) called', time_position)
-        self.backend.spotify.next_buffer_timestamp = time_position
+
+        if time_position == 0 and self._first_seek:
+            self._first_seek = False
+            logger.debug('Skipping seek due to issue #300')
+            return
+
+        self.backend.spotify.buffer_timestamp = audio.millisecond_to_clocktime(
+            time_position)
         self.backend.spotify.session.seek(time_position)