+++ mopidy/core/library.py
@@ -41,7 +41,7 @@ class LibraryController(object):
         query = query or kwargs
         futures = [
             b.library.find_exact(**query) for b in self.backends.with_library]
-        return pykka.get_all(futures)
+        return [result for result in pykka.get_all(futures) if result]
 
     def lookup(self, uri):
         """
@@ -101,4 +101,4 @@ class LibraryController(object):
         query = query or kwargs
         futures = [
             b.library.search(**query) for b in self.backends.with_library]
-        return pykka.get_all(futures)
+        return [result for result in pykka.get_all(futures) if result]